Fix dtype conversions.

+++ src/invoke_training/pipelines/_experimental/sd_dpo_lora/train.py
@@ -229,7 +229,10 @@ def train(config: SdDirectPreferenceOptimizationLoraConfig, callbacks: list[Pipe
 
     logger.info("Loading models.")
     tokenizer, noise_scheduler, text_encoder, vae, unet = load_models_sd(
-        model_name_or_path=config.model, hf_variant=config.hf_variant, base_embeddings=config.base_embeddings
+        model_name_or_path=config.model,
+        hf_variant=config.hf_variant,
+        base_embeddings=config.base_embeddings,
+        dtype=weight_dtype,
     )
     ref_text_encoder = copy.deepcopy(text_encoder)
     ref_unet = copy.deepcopy(unet)
